Add ParallelMap functionality to map.go
Added a new function, ParallelMap, to map.go to allow for parallel manipulation of a slice that transforms it to a different type. This was implemented to enhance performance by concurrently executing iterations, while maintaining the order of results. It also handles any possible errors which could occur during the execution.